split files up, test is separate
